+++ lib/simple_form/components/labels.rb
@@ -66,7 +66,7 @@ module SimpleForm
 
       # First check labels translation and then human attribute name.
       def label_translation #:nodoc:
-        if SimpleForm.translate_labels && (translated_label = translate(:labels))
+        if SimpleForm.translate_labels && (translated_label = translate_from_namespace(:labels))
           translated_label
         elsif object.class.respond_to?(:human_attribute_name)
           object.class.human_attribute_name(reflection_or_attribute_name.to_s)