+++ lib/simple_form/inputs/collection_input.rb
@@ -98,9 +98,15 @@ module SimpleForm
       end
 
       def translate_collection
-        if translated_collection = translate(:options)
+        if translated_collection = translate_from_namespace(:options)
           @collection = collection.map do |key|
-            [translated_collection[key] || key, key.to_s]
+            html_key = "#{key}_html".to_sym
+
+            if translated_collection[html_key]
+              [translated_collection[html_key].html_safe || key, key.to_s]
+            else
+              [translated_collection[key] || key, key.to_s]
+            end
           end
           true
         end
@@ -110,7 +116,7 @@ module SimpleForm
         if options[key] == :translate
           namespace = key.to_s.pluralize
 
-          options[key] = translate(namespace, true)
+          options[key] = translate_from_namespace(namespace, true)
         end
       end
     end