+++ test/inputs/collection_radio_buttons_input_test.rb
@@ -92,6 +92,34 @@ class CollectionRadioButtonsInputTest < ActionView::TestCase
     end
   end
 
+  test 'input should do automatic collection translation and preserve html markup' do
+    swap SimpleForm, boolean_style: :nested do
+      store_translations(:en, simple_form: { options: { user: {
+        gender: { male_html: '<strong>Male</strong>', female_html: '<strong>Female</strong>' }
+      } } } ) do
+        with_input_for @user, :gender, :radio_buttons, collection: [:male, :female]
+        assert_select 'input[type=radio][value=male]'
+        assert_select 'input[type=radio][value=female]'
+        assert_select 'label[for=user_gender_male]', 'Male'
+        assert_select 'label[for=user_gender_female]', 'Female'
+      end
+    end
+  end
+
+  test 'input should do automatic collection translation with keys prefixed with _html and a string value' do
+    swap SimpleForm, boolean_style: :nested do
+      store_translations(:en, simple_form: { options: { user: {
+        gender: { male_html: 'Male', female_html: 'Female' }
+      } } } ) do
+        with_input_for @user, :gender, :radio_buttons, collection: [:male, :female]
+        assert_select 'input[type=radio][value=male]'
+        assert_select 'input[type=radio][value=female]'
+        assert_select 'label[for=user_gender_male]', 'Male'
+        assert_select 'label[for=user_gender_female]', 'Female'
+      end
+    end
+  end
+
   test 'input should mark the current radio value by default' do
     @user.name = "Carlos"
     with_input_for @user, :name, :radio_buttons, collection: ['Jose', 'Carlos']